Breakup engine into multiple files
Initial attempt as splitting engine into multiple files. Better for long term work.
