+++ config.py
@@ -1,0 +1,26 @@
+config = {
+'DynRDSEnableRDS': '1',
+'DynRDSPSUpdateRate': '4',
+'DynRDSPSStyle': 'Merry|Christ-|  -mas!|{T}|{A}|[{N} of {C}]',
+'DynRDSRTUpdateRate': '8',
+'DynRDSRTSize': '32',
+'DynRDSRTStyle': 'Merry Christmas!|{T}[ by {A}]|[Track {N} of {C}]',
+'DynRDSPty': '2',
+'DynRDSPICode': '819b',
+'DynRDSTransmitter': 'None',
+'DynRDSFrequency': '100.1',
+'DynRDSPreemphasis': '75us',
+'DynRDSQN8066Gain': '0',
+'DynRDSQN8066SoftClipping': '0',
+'DynRDSQN8066AGC': '0',
+'DynRDSQN8066ChipPower': '113',
+'DynRDSQN8066PIHardwarePWM': 0,
+'DynRDSQN8066AmpPower': '0',
+'DynRDSStart': 'FPPDStart',
+'DynRDSStop': 'Never',
+'DynRDSCallbackLogLevel': 'INFO',
+'DynRDSEngineLogLevel': 'INFO',
+'DynRDSmpcEnable': '0',
+'DynRDSAdvPISoftwareI2C': '0',
+'DynRDSAdvPIPWMPin': '18,2'
+}
